Added the all products to the allproducts component. also updated som colours in tailwind config. Deleted a footer file not being used.
