+++ src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import App from "./App";
 import "./index.css";
-import store from "./components/store";
+import store from "./store/index";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
