+++ tailwind.config.cjs
@@ -18,11 +18,13 @@ module.exports = {
       },
       colors: {
         main: "#1A3D60",
-        secondary: "#F25F5C",
+        lightblue: "#4977A6",
+        secondary: "#E18886",
         accent: "#797D7F",
         background: "#F0F1F3",
         Gray: "#A7A2A2",
         text: "#333",
+        darkgreen: "#165664",
       },
     },
     plugins: [],
